requirements.txt.txt
# NexusMind v2 — Python Dependencies

# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Vector store
qdrant-client>=1.9.0

# HTTP
requests>=2.31.0
httpx>=0.27.0

# Document parsing
pdfplumber>=0.11.0
pypdf>=4.0.0
python-docx>=1.1.0

# Environment
python-dotenv>=1.0.0

# Data validation
pydantic>=2.7.0

# Evaluation / numerics
numpy>=1.26.0

# Utilities
tqdm>=4.66.0
python-multipart>=0.0.9   # FastAPI file uploads
